add permission hooks. change permissions link text to 'manage access' when user has admin role.
