+++ relay/components/AccessControl/IsRepoAdmin.tsx
@@ -1,24 +1,8 @@
-import {
-  FunctionComponent, useEffect, useState, useContext,
-} from 'react';
-import ValistContext from '../Valist/ValistContext';
+import { FunctionComponent } from 'react';
+import useRepoAdmin from '../../hooks/useRepoAdmin';
 
 const IsRepoAdmin:FunctionComponent<any> = (props) => {
-  const valist = useContext(ValistContext);
-  const [isRepoAdmin, setIsRepoAdmin] = useState(false);
-
-  useEffect(() => {
-    (async () => {
-      if (valist && valist.defaultAccount !== '0x0') {
-        try {
-          setIsRepoAdmin(await valist.isRepoAdmin(props.orgName, props.repoName, valist.defaultAccount));
-        } catch (e) {
-          setIsRepoAdmin(false);
-        }
-      }
-    })();
-  }, [valist]);
-
+  const isRepoAdmin = useRepoAdmin(props.orgName, props.repoName);
   if (isRepoAdmin) {
     return props.children;
   }
