+++ relay/components/AccessControl/ManageProjectCard.tsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import Link from 'next/link';
 import AddressIdenticon from '../Identicons/AddressIdenticon';
 import ValistContext from '../Valist/ValistContext';
+import useOrgAdmin from '../../hooks/useOrgAdmin';
 
 const User = ({ address }: { address: string }) => (
   <li className="py-4" key={address}>
@@ -23,6 +24,7 @@ const ManageProjectAccessCard = ({ orgName, projectName }: { orgName: string, pr
   const [repoAdmins, setRepoAdmins] = useState(['0x0']);
   const [repoDevs, setRepoDevs] = useState(['0x0']);
   const [orgOwners, setOrgOwners] = useState(['0x0']);
+  const isOrgAdmin = useOrgAdmin(orgName);
 
   const updateData = async () => {
     if (valist) {
@@ -66,7 +68,7 @@ const ManageProjectAccessCard = ({ orgName, projectName }: { orgName: string, pr
             <Link href={`/v/${orgName}/${projectName}/permissions`}>
               <a className="w-full flex justify-center items-center px-4 py-2 border border-gray-300 shadow-sm text-sm
               font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50">
-                View all
+                { isOrgAdmin ? 'Manage access' : 'View all' }
               </a>
             </Link>
           </div>
